test(org unit tree): test use-org-children hook
