+++ components/organisation-unit-tree/src/organisation-unit-node/use-org-data/use-org-data.test.js
@@ -25,14 +25,6 @@ describe('OrganisationUnitTree - useOrgData', () => {
                     id: 'A0000000000',
                     path: '/A0000000000',
                     displayName: 'Org Unit 1',
-                    children: [
-                        {
-                            id: 'A0000000001',
-                            path: '/A0000000000/A0000000001',
-                            children: [],
-                            displayName: 'Org Unit 2',
-                        },
-                    ],
                 }
             }
 
@@ -46,7 +38,7 @@ describe('OrganisationUnitTree - useOrgData', () => {
         </CustomDataProvider>
     )
 
-    it('should respond with `loading: false`, `error: null` and `data: { displayName, id }` initially', () => {
+    it('should respond with `loading: true`, `error: null` and `data: { displayName, id }` initially', () => {
         const { result } = renderHook(
             () => useOrgData('A0000000000', { displayName: 'Display name' }),
             { wrapper }
@@ -74,14 +66,6 @@ describe('OrganisationUnitTree - useOrgData', () => {
                 id: 'A0000000000',
                 path: '/A0000000000',
                 displayName: 'Org Unit 1',
-                children: [
-                    {
-                        id: 'A0000000001',
-                        path: '/A0000000000/A0000000001',
-                        children: [],
-                        displayName: 'Org Unit 2',
-                    },
-                ],
             },
         })
     })
@@ -110,32 +94,6 @@ describe('OrganisationUnitTree - useOrgData', () => {
             loading: false,
             error: new Error('Error message'),
             data: { displayName: 'Display name', id: 'A0000000000' },
-        })
-    })
-
-    it('should call the onComplete callback', async () => {
-        const onComplete = jest.fn()
-        const options = { onComplete, displayName: 'Display name' }
-
-        const { waitForNextUpdate } = renderHook(
-            () => useOrgData('A0000000000', options),
-            { wrapper }
-        )
-
-        await waitForNextUpdate()
-
-        expect(onComplete).toHaveBeenCalledWith({
-            id: 'A0000000000',
-            path: '/A0000000000',
-            displayName: 'Org Unit 1',
-            children: [
-                {
-                    id: 'A0000000001',
-                    path: '/A0000000000/A0000000001',
-                    children: [],
-                    displayName: 'Org Unit 2',
-                },
-            ],
         })
     })
 
@@ -182,143 +140,5 @@ describe('OrganisationUnitTree - useOrgData', () => {
             }),
             expect.objectContaining({}) // contains the `signal`
         )
-    })
-
-    it("should sort the node's children alphabetically by default", async () => {
-        const dataProviderDataWithUnsortedChildren = {
-            organisationUnits: jest.fn((type, { id }) => {
-                if (id === 'A0000000000') {
-                    return {
-                        id: 'A0000000000',
-                        path: '/A0000000000',
-                        displayName: 'Org Unit 1',
-                        children: [
-                            {
-                                id: 'A0000000002',
-                                path: '/A0000000000/A0000000002',
-                                children: [],
-                                displayName: 'Org Unit 3',
-                            },
-                            {
-                                id: 'A0000000001',
-                                path: '/A0000000000/A0000000001',
-                                children: [],
-                                displayName: 'Org Unit 2',
-                            },
-                        ],
-                    }
-                }
-
-                return Promise.reject(`No org unit with id "${id}"`)
-            }),
-        }
-
-        const wrapperWithUnsortedChildren = ({ children }) => (
-            <CustomDataProvider data={dataProviderDataWithUnsortedChildren}>
-                {children}
-            </CustomDataProvider>
-        )
-
-        const { result, waitForNextUpdate } = renderHook(
-            () => useOrgData('A0000000000', { displayName: 'Display name' }),
-            { wrapper: wrapperWithUnsortedChildren }
-        )
-
-        await waitForNextUpdate()
-
-        expect(result.current).toEqual({
-            loading: false,
-            error: null,
-            data: {
-                id: 'A0000000000',
-                path: '/A0000000000',
-                displayName: 'Org Unit 1',
-                children: [
-                    {
-                        id: 'A0000000001',
-                        path: '/A0000000000/A0000000001',
-                        children: [],
-                        displayName: 'Org Unit 2',
-                    },
-                    {
-                        id: 'A0000000002',
-                        path: '/A0000000000/A0000000002',
-                        children: [],
-                        displayName: 'Org Unit 3',
-                    },
-                ],
-            },
-        })
-    })
-
-    it(`should not sort the node's children alphabetically when "suppressAlphabeticalSorting" is true`, async () => {
-        const dataProviderDataWithUnsortedChildren = {
-            organisationUnits: jest.fn((type, { id }) => {
-                if (id === 'A0000000000') {
-                    return {
-                        id: 'A0000000000',
-                        path: '/A0000000000',
-                        displayName: 'Org Unit 1',
-                        children: [
-                            {
-                                id: 'A0000000002',
-                                path: '/A0000000000/A0000000002',
-                                children: [],
-                                displayName: 'Org Unit 3',
-                            },
-                            {
-                                id: 'A0000000001',
-                                path: '/A0000000000/A0000000001',
-                                children: [],
-                                displayName: 'Org Unit 2',
-                            },
-                        ],
-                    }
-                }
-
-                return Promise.reject(`No org unit with id "${id}"`)
-            }),
-        }
-
-        const wrapperWithUnsortedChildren = ({ children }) => (
-            <CustomDataProvider data={dataProviderDataWithUnsortedChildren}>
-                {children}
-            </CustomDataProvider>
-        )
-
-        const options = {
-            displayName: 'Display name',
-            suppressAlphabeticalSorting: true,
-        }
-        const { result, waitForNextUpdate } = renderHook(
-            () => useOrgData('A0000000000', options),
-            { wrapper: wrapperWithUnsortedChildren }
-        )
-
-        await waitForNextUpdate()
-
-        expect(result.current).toEqual({
-            loading: false,
-            error: null,
-            data: {
-                id: 'A0000000000',
-                path: '/A0000000000',
-                displayName: 'Org Unit 1',
-                children: [
-                    {
-                        id: 'A0000000002',
-                        path: '/A0000000000/A0000000002',
-                        children: [],
-                        displayName: 'Org Unit 3',
-                    },
-                    {
-                        id: 'A0000000001',
-                        path: '/A0000000000/A0000000001',
-                        children: [],
-                        displayName: 'Org Unit 2',
-                    },
-                ],
-            },
-        })
     })
 })
